+++ core/lib/pakyow/server.rb
@@ -12,6 +12,24 @@ module Pakyow
 
     def initialize(context, endpoint:, protocol:, scheme:)
       super(context, endpoint, protocol, scheme)
+
+      @server = nil
+    end
+
+    def run
+      @endpoint.bind do |server|
+        @server = server
+        @server.listen(Socket::SOMAXCONN)
+        @server.accept_each(&method(:accept))
+      rescue Async::Wrapper::Cancelled
+        # the endpoint was closed
+      end
+
+      self
+    end
+
+    def shutdown
+      @server&.io&.close
     end
   end
 end